+++ R/export-methods.R
@@ -10,13 +10,13 @@
 #'
 #' @examples
 #' object <- S4Vectors::DataFrame(
-#'     transcriptID = c(
+#'     "transcriptId" = c(
 #'         "tx0001",
 #'         "tx0002",
 #'         "tx0003",
 #'         "tx0004"
 #'     ),
-#'     geneID = c(
+#'     "geneId" = c(
 #'         "gene0001",
 #'         "gene0001",
 #'         "gene0002",